use-cases/churn-prediction/bigquery/requirements.txt
google-cloud-bigquery>=2.30.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
numpy>=1.20.0
google-cloud-storage>=1.42.0
db-dtypes>=1.0.0
pyarrow>=5.0.0